Rename ebgp.utils plugin to bgp.session
