+++ netsim/extra/bgp.session/eos.j2
