+++ netsim/extra/bgp.session/routeros7.j2
