+++ netsim/extra/bgp.session/sros.j2
