+++ netsim/extra/bgp.session/vyos.j2
